# Conflicts:
#	Anytype/Sources/Models/Documents/Events/Subscriptions/Internals/SubscriptionToggler.swift
#	Anytype/Sources/Models/Documents/Events/Subscriptions/SubscriptionIds.swift
#	Anytype/Sources/PresentationLayer/Home/CellDataBuilder/HomeCellDataBuilder.swift
#	Anytype/Sources/PresentationLayer/TextEditor/Assembly/EditorPageAssembly.swift
#	Anytype/Sources/PresentationLayer/TextEditor/Set/EditorSetViewModel.swift
#	Anytype/Sources/PresentationLayer/TextEditor/Set/Views/Popups/Filters/List/SetFiltersListViewModel.swift
#	Anytype/Sources/PresentationLayer/TextEditor/Set/Views/Popups/Sorts/List/SetSortsListViewModel.swift
#	Dependencies/Middleware/Lib.xcframework/Info.plist

+++ Anytype/Sources/PresentationLayer/Debug/FeatureFlags/DebugMenu.swift
@@ -32,7 +32,7 @@ struct DebugMenu: View {
         }
         .padding(20)
         .onChange(of: rowsPerPageInSet) { count in
-            guard let count = Int64(count) else { return }
+            guard let count = Int(count) else { return }
             UserDefaultsConfig.rowsPerPageInSet = count
         }
     }
